[TabsManager] Rewrite VsHierarchyWalker -> VsHierarchy
